# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "http server api docs",
        "title": "lol api docs",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/result": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "submit information to create result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "result"
                ],
                "summary": "create result",
                "parameters": [
                    {
                        "description": "result information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateResultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateResultReply"
                        }
                    }
                }
            }
        },
        "/api/v1/result/list": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list of results by paging and conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "result"
                ],
                "summary": "list of results by query parameters",
                "parameters": [
                    {
                        "description": "query parameters",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Params"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListResultsReply"
                        }
                    }
                }
            }
        },
        "/api/v1/result/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get result detail by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "result"
                ],
                "summary": "get result detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetResultByIDReply"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update result information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "result"
                ],
                "summary": "update result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "result information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateResultByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateResultByIDReply"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete result by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "result"
                ],
                "summary": "delete result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteResultByIDReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Column": {
            "type": "object",
            "properties": {
                "exp": {
                    "description": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like",
                    "type": "string"
                },
                "logic": {
                    "description": "logical type, default value is \"and\", support \u0026, and, ||, or",
                    "type": "string"
                },
                "name": {
                    "description": "column name",
                    "type": "string"
                },
                "value": {
                    "description": "column value"
                }
            }
        },
...
        "types.CreateResultReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "id": {
                            "description": "id",
                            "type": "integer"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.CreateResultRequest": {
            "type": "object",
            "properties": {
                "amountTotal": {
                    "description": "合计",
                    "type": "number"
                },
                "attach": {
                    "type": "string"
                },
                "bankType": {
                    "description": "银行类型",
                    "type": "string"
                },
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "eventType": {
                    "description": "事件类型",
                    "type": "string"
                },
                "outTradeNo": {
                    "description": "订单号",
                    "type": "string"
                },
                "payer": {
                    "description": "支付人",
                    "type": "string"
                },
                "resourceAppid": {
                    "description": "来源ID",
                    "type": "string"
                },
                "resourceMchid": {
                    "description": "来源商户ID",
                    "type": "string"
                },
                "successTime": {
                    "description": "成功时间",
                    "type": "string"
                },
                "tradeState": {
                    "description": "交易状态",
                    "type": "string"
                },
                "tradeStateDesc": {
                    "description": "交易状态描述",
                    "type": "string"
                },
                "tradeType": {
                    "description": "交易类型",
                    "type": "string"
                },
                "transactionID": {
                    "description": "交易ID",
                    "type": "string"
                }
            }
        },
...
        "types.DeleteResultByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.GetResultByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "result": {
                            "$ref": "#/definitions/types.ResultObjDetail"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.ListResultsReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "results": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.ResultObjDetail"
                            }
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.Params": {
            "type": "object",
            "properties": {
                "columns": {
                    "description": "query conditions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Column"
                    }
                },
                "limit": {
                    "description": "lines per page",
                    "type": "integer"
                },
                "page": {
                    "description": "page number, starting from page 0",
                    "type": "integer"
                },
                "sort": {
                    "description": "sorted fields, multi-column sorting separated by commas",
                    "type": "string"
                }
            }
        },
...
        "types.ResultObjDetail": {
            "type": "object",
            "properties": {
                "amountTotal": {
                    "description": "合计",
                    "type": "number"
                },
                "attach": {
                    "type": "string"
                },
                "bankType": {
                    "description": "银行类型",
                    "type": "string"
                },
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "eventType": {
                    "description": "序号",
                    "type": "string"
                },
                "id": {
                    "description": "convert to uint64 id",
                    "type": "integer"
                },
                "outTradeNo": {
                    "description": "订单号",
                    "type": "string"
                },
                "payer": {
                    "description": "支付人",
                    "type": "string"
                },
                "resourceAppid": {
                    "description": "来源ID",
                    "type": "string"
                },
                "resourceMchid": {
                    "description": "来源商户ID",
                    "type": "string"
                },
                "successTime": {
                    "description": "成功时间",
                    "type": "string"
                },
                "tradeState": {
                    "description": "交易状态",
                    "type": "string"
                },
                "tradeStateDesc": {
                    "description": "交易状态描述",
                    "type": "string"
                },
                "tradeType": {
                    "description": "交易类型",
                    "type": "string"
                },
                "transactionID": {
                    "description": "交易ID",
                    "type": "string"
                }
            }
        },
...
        "types.UpdateResultByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.UpdateResultByIDRequest": {
            "type": "object",
            "properties": {
                "amountTotal": {
                    "description": "合计",
                    "type": "number"
                },
                "attach": {
                    "type": "string"
                },
                "bankType": {
                    "description": "银行类型",
                    "type": "string"
                },
                "createAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "eventType": {
                    "description": "序号",
                    "type": "string"
                },
                "id": {
                    "description": "uint64 id",
                    "type": "integer"
                },
                "outTradeNo": {
                    "description": "订单号",
                    "type": "string"
                },
                "payer": {
                    "description": "支付人",
                    "type": "string"
                },
                "resourceAppid": {
                    "description": "来源ID",
                    "type": "string"
                },
                "resourceMchid": {
                    "description": "来源商户ID",
                    "type": "string"
                },
                "successTime": {
                    "description": "成功时间",
                    "type": "string"
                },
                "tradeState": {
                    "description": "交易状态",
                    "type": "string"
                },
                "tradeStateDesc": {
                    "description": "交易状态描述",
                    "type": "string"
                },
                "tradeType": {
                    "description": "交易类型",
                    "type": "string"
                },
                "transactionID": {
                    "description": "交易ID",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type Bearer your-jwt-token to Value",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}